fix: conditionally load katex css detecting if the page uses katex
