+++ apps/client/src/app/(others)/post/[slug]/@katex/page.tsx
@@ -1,0 +1,5 @@
+import 'katex/dist/katex.min.css';
+
+export default function KatexPage() {
+  return null;
+}
