--- apps/client/src/features/post/katex.tsx
@@ -1,6 +1,0 @@
-'use client';
-import 'katex/dist/katex.min.css';
-
-export default function KatexCssDummy() {
-  return null;
-}
